# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Witme API Documentation",
        "description": "WitMe API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Логин",
                "description": "Логин",
                "operationId": "a7997e7c1e3bfed64d4e56de82415931",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LoginUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMjA3LjE1NC4yMTYuMTQvYXBpL3YxL2F1dGgvcmVnaXN0ZXIiLCJpYXQiOjE3MjgxNjA2ODIsImV4cCI6MTcyODE2NjY4MiwibmJmIjoxNzI4MTYwNjgyLCJqdGkiOiJFeXJKb05yNXhFdWZpaVV5Iiwic3ViIjoiOSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.NkCVbpCtawv5lAjXmVLcRdwzYgpmhJ1FbhV1acSRKEE"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "6000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Регистрация",
                "description": "Регистрация",
                "operationId": "b9e746d77661fb8df55d0ac63e1ece77",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RegisterUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMjA3LjE1NC4yMTYuMTQvYXBpL3YxL2F1dGgvcmVnaXN0ZXIiLCJpYXQiOjE3MjgxNjA2ODIsImV4cCI6MTcyODE2NjY4MiwibmJmIjoxNzI4MTYwNjgyLCJqdGkiOiJFeXJKb05yNXhFdWZpaVV5Iiwic3ViIjoiOSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.NkCVbpCtawv5lAjXmVLcRdwzYgpmhJ1FbhV1acSRKEE"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "6000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Выйти",
                "description": "Выйти",
                "operationId": "f6b8957b4c0b25636613a9f0512a8ba9",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/auth/refresh": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Рефреш",
                "description": "Рефреш",
                "operationId": "c42c3cd7fd7fe0dbd5bcc42962059e90",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access_token": {
                                            "type": "string",
                                            "example": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vMjA3LjE1NC4yMTYuMTQvYXBpL3YxL2F1dGgvcmVnaXN0ZXIiLCJpYXQiOjE3MjgxNjA2ODIsImV4cCI6MTcyODE2NjY4MiwibmJmIjoxNzI4MTYwNjgyLCJqdGkiOiJFeXJKb05yNXhFdWZpaVV5Iiwic3ViIjoiOSIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.NkCVbpCtawv5lAjXmVLcRdwzYgpmhJ1FbhV1acSRKEE"
                                        },
                                        "expires_in": {
                                            "type": "integer",
                                            "example": "6000"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Cписок книг",
                "description": "Получить список книг",
                "operationId": "61b0fdd35b8c14afc6e29eb90598649a",
                "responses": {
                    "200": {
                        "description": "Cписок книг",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BooksResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Создать книгу",
                "description": "Создать книгу",
                "operationId": "5df8eb3d652dcc56b60cd15be3cbbf85",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/books/{id}": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Cписок книг",
                "description": "Получить список книг",
                "operationId": "c5b00bf4246c67b18e3092bd3665ea29",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Cписок книг",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/BookResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Редактировать книгу",
                "description": "Редактировать книгу",
                "operationId": "e9e7186a14d97ce57f6f581ad4a475bd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Books"
                ],
                "summary": "Удалить книгу",
                "description": "Удалить книгу",
                "operationId": "0d28c2ad15a8fdbe85c2c95806b007f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/profile/update": {
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Обновить профиль",
                "description": "Обновить профиль",
                "operationId": "93b74270044df24a0d6db14b1bd076d5",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/books/{bookId}/sessions": {
            "post": {
                "tags": [
                    "Book Sessions"
                ],
                "summary": "Создать сессию книги",
                "description": "Создать сессию книги",
                "operationId": "store",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookSessionRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "422": {
                        "description": "Unprocessable Content",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Something went wrong"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "StoreBookRequest": {
                "required": [
                    "name",
                    "author",
                    "book_type_id",
                    "book_status_id"
                ],
                "properties": {
                    "name": {
                        "description": "Название книги",
                        "type": "string"
                    },
                    "author": {
                        "description": "Имя автора",
                        "type": "string"
                    },
                    "total_duration": {
                        "description": "Длительность/Кол-во страниц",
                        "type": "string"
                    },
                    "description": {
                        "description": "Описание книги",
                        "type": "string"
                    },
                    "image": {
                        "description": "Фото обложки",
                        "type": "binary"
                    },
                    "book_type_id": {
                        "description": "Тип книги",
                        "type": "integer"
                    },
                    "book_status_id": {
                        "description": "Статус книги",
                        "type": "integer"
                    },
                    "planning_date": {
                        "description": "Дата планируемого читания",
                        "type": "string"
                    },
                    "start_date": {
                        "description": "Дата начала читания",
                        "type": "string"
                    },
                    "end_date": {
                        "description": "Дата конца читания",
                        "type": "string"
                    },
                    "rating": {
                        "description": "Оценка книги",
                        "type": "number",
                        "format": "float"
                    },
                    "review": {
                        "description": "Отзыв",
                        "type": "string"
                    },
                    "avg_emoji": {
                        "description": "Эмодзи",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "UpdateBookRequest": {
                "properties": {
                    "name": {
                        "description": "Название книги",
                        "type": "string"
                    },
                    "author": {
                        "description": "Автор книги",
                        "type": "string"
                    },
                    "total_duration": {
                        "description": "Длительность/кол-во книги",
                        "type": "integer"
                    },
                    "description": {
                        "description": "Описание книги",
                        "type": "string"
                    },
                    "image": {
                        "description": "Обложка книги",
                        "type": "binary"
                    },
                    "book_type_id": {
                        "description": "ID тип книги",
                        "type": "integer"
                    },
                    "book_status_id": {
                        "description": "ID статус книги",
                        "type": "integer"
                    },
                    "planning_date": {
                        "description": "Планируемая дата чтения",
                        "type": "string"
                    },
                    "start_date": {
                        "description": "Дата начала чтения",
                        "type": "string"
                    },
                    "end_date": {
                        "description": "Дата конец чтения",
                        "type": "integer"
                    },
                    "rating": {
                        "description": "Оценка",
                        "type": "number",
                        "format": "float"
                    },
                    "review": {
                        "description": "Отзыв",
                        "type": "string"
                    },
                    "avg_emoji": {
                        "description": "Эмодзи",
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "StoreBookSessionRequest": {
                "required": [
                    "session_duration"
                ],
                "properties": {
                    "session_duration": {
                        "description": "Длительность сессии (в секундах)",
                        "type": "integer"
                    },
                    "current_duration": {
                        "description": "Страниц прочитано/Секунд прослушано книги",
                        "type": "integer"
                    },
                    "notes": {
                        "description": "Заметки сесии книги",
                        "type": "array",
                        "items": {
                            "properties": {
                                "comment": {
                                    "type": "string",
                                    "example": "Sigma"
                                },
                                "files": {
                                    "type": "array",
                                    "items": {
                                        "type": "file",
                                        "example": "photo1.jpg"
                                    }
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
            "LoginUserRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Почта",
                        "type": "string"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "RegisterUserRequest": {
                "required": [
                    "email",
                    "password"
                ],
                "properties": {
                    "email": {
                        "description": "Почта",
                        "type": "string"
                    },
                    "password": {
                        "description": "Пароль",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdateUserRequest": {
                "properties": {
                    "name": {
                        "description": "Имя",
                        "type": "string"
                    },
                    "avatar": {
                        "description": "Аватар",
                        "type": "binary"
                    }
                },
                "type": "object"
            },
            "BookResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    },
                    "author": {
                        "type": "string"
                    },
                    "book_type": {
                        "type": "string"
                    },
                    "book_status": {
                        "type": "string"
                    },
                    "total_duration": {
                        "type": "integer"
                    },
                    "notes_amount": {
                        "type": "integer"
                    },
                    "notes": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NoteResource"
                        }
                    },
                    "description": {
                        "type": "string"
                    },
                    "image": {
                        "type": "string"
                    },
                    "planning_date": {
                        "type": "string"
                    },
                    "start_date": {
                        "type": "string"
                    },
                    "end_date": {
                        "type": "integer"
                    },
                    "rating": {
                        "type": "number",
                        "format": "float"
                    },
                    "review": {
                        "type": "string"
                    },
                    "avg_emoji": {
                        "type": "integer"
                    }
                },
                "type": "object"
            },
            "BooksResource": {
                "type": "array",
                "items": {
                    "$ref": "#/components/schemas/BookResource"
                }
            },
            "FileResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "path": {
                        "type": "string"
                    },
                    "original_name": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "NoteResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "session_id": {
                        "type": "string"
                    },
                    "comment": {
                        "type": "string"
                    },
                    "created_at": {
                        "type": "string"
                    },
                    "files": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/FileResource"
                        }
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "Authentication"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Profile",
            "description": "Profile"
        },
        {
            "name": "Book Sessions",
            "description": "Book Sessions"
        }
    ]
}